V8
Se ha cambiado el timer0 para que siempre cuente
y para calcular se compara con lo que lleva contado

Falta probar la maquina de estados de los botones

+++ practica2/timer2.h
@@ -7,7 +7,7 @@
 
 #ifndef _TIMER2_H_
 #define _TIMER2_H_
-
+#include <inttypes.h>
 /*--- variables visibles del módulo timer.c/timer.h ---*/
 int timer2_num_int;
 int pruebaTimer2;
@@ -34,6 +34,6 @@ void timer2_empezar();
  * lee la cuenta actual del temporizador y el número de interrupciones generadas, y
  * devuelve el tiempo transcurrido en microsegundos.
  */
-int timer2_leer();
+double timer2_leer();
 
 #endif /* _TIMER_H_ */
